separated jsearch models
I hate codefactor
