+++ DDEyC_API/DataAccess/Models/JSearch/JSearchJobHighlights.cs
@@ -1,0 +1,9 @@
+namespace DDEyC_API.Models.JSearch{
+    
+    public class JSearchJobHighlights
+    {
+        public List<string> Qualifications { get; set; } = new();
+        public List<string> Responsibilities { get; set; } = new();
+        public List<string> Benefits { get; set; } = new();
+    }
+}
