+++ DDEyC_API/DataAccess/Models/JSearch/JSearchParameters.cs
@@ -1,0 +1,11 @@
+namespace DDEyC_API.Models.JSearch
+{
+    public class JSearchParameters
+    {
+        public string query { get; set; }
+        public int page { get; set; }
+        public int num_pages { get; set; }
+        public string date_posted { get; set; }
+    }
+
+}
