+++ DDEyC_API/DataAccess/Models/JSearch/JSearchResponse.cs
@@ -1,0 +1,9 @@
+namespace DDEyC_API.Models.JSearch{
+     public class JSearchResponse
+    {
+        public string status { get; set; }
+        public string request_id { get; set; }
+        public JSearchParameters parameters { get; set; }
+        public List<JSearchJob> data { get; set; }
+    }
+}
